New fire capability, tile drainage and SOM accumulation
This version has the vertically resolved heat/mass transfer in the
canopy for dealing with explicit fire ignition and combustion process.
Also updated the water table setup and output for tile drainage. The
numerical iteration cycles are also revised. Accordingly, solver setup
in the example inputs are also modified. The treatment for SOM accumulation
due to various processes are also improved.